Move CheckoutContextInterface into types file
