+++ packages/composite-checkout/src/lib/checkout-context.ts
@@ -1,25 +1,5 @@
 import { createContext } from 'react';
-import {
-	StepChangedCallback,
-	CheckoutPageErrorCallback,
-	PaymentMethod,
-	PaymentProcessorProp,
-	TransactionStatusManager,
-	PaymentMethodChangedCallback,
-} from '../types';
-
-export interface CheckoutContextInterface {
-	allPaymentMethods: PaymentMethod[];
-	disabledPaymentMethodIds: string[];
-	setDisabledPaymentMethodIds: ( methods: string[] ) => void;
-	paymentMethodId: string | null;
-	setPaymentMethodId: ( id: string ) => void;
-	transactionStatusManager: TransactionStatusManager | null;
-	paymentProcessors: PaymentProcessorProp;
-	onPageLoadError?: CheckoutPageErrorCallback;
-	onStepChanged?: StepChangedCallback;
-	onPaymentMethodChanged?: PaymentMethodChangedCallback;
-}
+import { CheckoutContextInterface } from '../types';
 
 const defaultCheckoutContext: CheckoutContextInterface = {
 	allPaymentMethods: [],
